refactor(tests): fix eslint errors
