+++ src/components/Table/Table.test.helpers.js
@@ -1,0 +1,25 @@
+export const mockActions = {
+  pagination: {
+    onChangePage: jest.fn(),
+  },
+  toolbar: {
+    onApplyFilter: jest.fn(),
+    onToggleFilter: jest.fn(),
+    onToggleColumnSelection: jest.fn(),
+    onClearAllFilters: jest.fn(),
+    onCancelBatchAction: jest.fn(),
+    onApplyBatchAction: jest.fn(),
+    onApplySearch: jest.fn(),
+    onDownloadCSV: jest.fn(),
+  },
+  table: {
+    onRowSelected: jest.fn(),
+    onRowClicked: jest.fn(),
+    onRowExpanded: jest.fn(),
+    onSelectAll: jest.fn(),
+    onChangeSort: jest.fn(),
+    onApplyRowAction: jest.fn(),
+    onEmptyStateAction: jest.fn(),
+    onChangeOrdering: jest.fn(),
+  },
+};
